add deploy yaml
