+++ deploy/webhook.yaml
@@ -1,0 +1,45 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: pvc-accessor-deployment
+  namespace: default # NOTE: change the namespace
+  labels:
+    app: pvc-accessor
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: pvc-accessor
+  template:
+    metadata:
+      labels:
+        app: pvc-accessor
+    spec:
+      containers:
+      - name: pvc-accessor
+        image: f10atin9/webhook-accessor:v1.0 # change the image if you wish to use your own custom validation server image
+        imagePullPolicy: IfNotPresent
+        args: ['--tls-cert-file=/etc/pvc-accessor-webhook/certs/cert.pem', '--tls-private-key-file=/etc/pvc-accessor-webhook/certs/key.pem']
+        ports:
+        - containerPort: 443 # change the port as needed
+        volumeMounts:
+          - name: pvc-accessor-webhook-certs
+            mountPath: /etc/pvc-accessor-webhook/certs
+            readOnly: true
+      volumes:
+        - name: pvc-accessor-webhook-certs
+          secret:
+            secretName: pvc-validation-secret
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: pvc-accessor-service
+  namespace: default # NOTE: change the namespace
+spec:
+  selector:
+    app: pvc-accessor
+  ports:
+    - protocol: TCP
+      port: 443 # Change if needed
+      targetPort: 443 # Change if the webserver image expects a different port
